producer/requirements.txt
functions-framework
google-cloud-pubsub
streamlit
